cores/VideoPlayer/ProcessInfo: remove the override system
In https://github.com/xbmc/xbmc/pull/9390, FernetMenta explained that
this override system is the design of choice in Kodi, because it
avoids "cluttering the code with ifdefs".

So this commit is a proof-of-concept which demonstrates that the
opposite is true.

In the "core" source file ProcessInfo.cpp, it moves one #ifdef to
another location.  In return, it eliminates the source file
ProcessInfoPi.cpp and its header completely, which contained another
#ifdef.  The autotools/cmake scripts are simplified by a big margin
because there is no fragile build-time detection of override source
files using shell wildcards.

So what we get: lots of complicated and fragile code eliminates, and
number of #ifdefs reduced.  The new code is easily readable and
understandable.

See commit 4a9d777fe296d79c4cc40774c1c15254021dfe31 for what can
happen when the code is hard to understand.